overlord,tests: perform soft refresh check in doInstall
When a snap is installed or refreshed, perform a soft check. That is,
see if there are any applications running that would interfere with the
refresh process.

The soft check is the first half of the feature: the hard check is next,
and verifies that the applications are still not running just before
unlinking the old snap revision.

Note that the information about the snap to be refreshed needs to be
loaded from the state because we need to know the current set of apps
and hooks that are there.

This patch also contains a simple test for the soft refresh check of the
app-refresh-awareness feature. When the feature is enabled refreshing an
application while it is running is rejected.

Two tests that were not locking the state were amended to do so since we
now need to inspect the experimental flags in doInstall.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>