tests: run create-user on core devices
The current "snap create-user" test runs on classic only. This
is bad for multiple reasons. The main use-case of this feature
are core devices and we are considering to drop support for this
on classic entirely.

So this PR changes the test to run on core16 now and once the
"shadow" package SRU pending in
https://launchpad.net/ubuntu/+source/shadow/1:4.5-1ubuntu2
is merged we can enable it on core18 too.