overlord/snapstate: inhibit refresh for up to a week
When a snap refresh is inhibited by running applications or hooks the
time of that event is recorded in the snap state. When a snap
application process is inhibiting a refresh from occurring for over one
week then force-refresh the snap anyway.

The actual value is not precisely agreed upon yet. Perhaps it should be
configurable. Perhaps it should be constant but different.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>