cmd/snap-seccomp-blacklist: explain syscall resolution
Interestingly, libseccomp has a mechanis to resolve a system call
number, that is absent from the current architecture, to a virtual
system call number that can be used internally by libseccomp.

This allows creating rules for system calls for other architectures than
the native one. In addition it allows for special-casing of multiplexing
system calls that are split into dedicated new system calls, like the
socket system call was.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>