cmd/libsnap: rework mountinfo parser
This patch changes the parser from a for loop with complex conditions
and switch statement to a endless for loop where each termination
condition is encoded explicitly.

This allows us to behave consistently with parsing the terminating NUL
byte and the separating space, for which there were previously some
concerns.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>