Wait for man db cache is updated before do daemon-reexec after install snpad
The service run-<id>.service which runs systemctl start
man-db-cache-update fails when we do daemon-reexec after install snpad,
leaving the system in degraded state. This is happening with the new
images for fedora.

To test this error and the fix use:
- fedora-29-64:
    image: test-fedora-1

The idea of the fix is to wait until this service finishes and then make
daemon-reexec.