cmd/snap-update-ns: rename ctx to upCtx
While doing a code review for the original branch Pedronis and me have
agreed to use upCtx to differentiate the plain ctx from a golang context
type.

Since the intermediate patches used ctx to avoid churn and re-basing we
need one follow up patch to rename everything at once. This is it.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>