devicestate: set "new-model" on the remodel change
The old code was setting the "new-model" data only to the "set-model"
task. When we move to remodel to new kernels/bases we need this
data available more broadly, i.e. the install code needs to check
if the new kernel can be installed. This PR adds the new model
directly to the remodel change.