interfaces/docker-support: add overlayfs accesses for ubuntu core
On Ubuntu Core, because of an apparmor bug with how the target dirs are
resolved with overlafys mounts, the target mount dir resolves to the
underlying bind mount from the /system-data directory, which was not
allowed in the default AppArmor policy.

This conditionally adds access to
/systemd-data/{$SNAP_COMMON,$SNAP_DATA}/** if we are on Ubuntu Core so
that the overlay2 storage-driver may be used with the docker snap.

The current default storage-driver for the docker snap is AUFS, which
has been deprecated upstream and will not be supported at some point in
the near future, so we need to prepare to switch to only supporting
overlay2, and as such need to be able to do overlayfs mounts on Ubuntu
Core.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>