snapstate: revert "overlord/snapstate: remove PlugsOnly"
This reverts commit 76388600020ff8dea740ddcd5a3ad1030e04ce2d.

It turns out that we removed PlugsOnly a bit too early. While
it is not used right now it will be used in the future by a
PR that Samuele is working on. So revert the removal here
to ensure it does not slow the work on this PR down.