store,daemon,client: allow user-agent setting for Search()
This will take the first "product" from the user-agent that talks
to snapd and adds it as an extra user-agent when talking to the
store. The use-case for this is gnome-software which can set
the agent to "gnome-software" so that the store can gather data
about how many CLI vs GUI users we have.

This initial PR implements it just for "Search" but if this
approach is acceptable it will be added for Install as well.