snapstate: auto-install snapd when needed
* snapstate: auto-install snapd when needed

When installing a snap that does not require "core" on a fresh
system, currently we do nothing special. This means however that
on such systems there is no "core" or "snapd" snap installed.

This means that no system interfaces are available. This PR
fixes this by auto-installing the snapd snap if its missing.

As a side effect, "snap install snapd" on classic is no longer
blocked (it still is on core).

* snapstate: add more tests for snapd auto-install with core18/other-bases

* snapstate: rework the check if the snapd snap can be installed on a Core system

* snapstate: use isInstalled instead of {snapd,core}SnapInstalled

* snapstate: update doInstall now that the "snapd" install check moved to checkInstallPreconditions