overlord/hookstate: don't run handler unless hooksup.Always
Without this change, a non-hijacked hook in a snap that had no script
file would still result in the handler getting called. This change
avoids that: unless the (new) Always flag is set, the hook runner
bails early. This means we avoid a bunch of work (e.g. setting up the
context) for hooks that are going to no-op anyway, while avoiding (for
now) having to pass a flag to Done() to specify whether the script
actually ran. This last bit is important for health checks, and is the
driver for this change.