cmd/snap: allow option descriptions to start with the command
Before this change, option descriptions that were slightly
self-referential and started with the command name would trigger the
linter warning:

    2018/12/12 19:51:16.123791 main.go:152: description of advise-snap's "from-apt" is lowercase: "advise-snap vil tale med apt via en apt-hook"

this change makes it so that an option description can start with the
command without raising a warning.

It also fixes that particular option to be more correct, and
discourages this particular translation.