gadget: drop excessive offset-write validation
The offset-write property defines the position *within the volume* at which, the
offset of the current element will be written at. A structure-relative
offset-write may not be located within a structure it refers to, and there is no
requirement for the pointer value to be within the referred structure size.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>