systemd: workaround older systemctl show quirks
Older version of systemctl show --property=.. would exit with non-0 status when
asked for properties that are not defined for given type of unit. For example,
--property=Type fails for units of socket or timer type. On CentOS 7.6, snap
services would then drop an error like so:

$ snap services
error: cannot get status of services of app "daemon": [show
       --property=Id,ActiveState,UnitFileState,Type snap.lxd.daemon.service
       snap.lxd.daemon.unix.socket] failed with exit status 1: Type=simple
       Id=snap.lxd.daemon.service ActiveState=inactive UnitFileState=static

       Id=snap.lxd.daemon.unix.socket ActiveState=active UnitFileState=enabled

Workaround that by splitting how we ask for properties of systemd units which is
compatible with old and new systemd.

Fixes: https://bugs.launchpad.net/bugs/1828847

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>