daemon: refactor user ops to api_users
This is a pure refactor: it moves login/logout/create user/list users
out to api_users.go.  It also moves the test for create/list to
api_users_test.go; the moving of the login/logout tests is left for
later (it's fiddly and intertwined), as is making the tests sit in the
daemon_test package.