cmd/snap-confine: don't pass MS_SLAVE along with MS_BIND
While making other changes I noticed that we have a code sequence
performing bind mount and switching it to slave propagation mode:

    sc_do_mount(SNAP_MOUNT_DIR, dst, NULL, MS_BIND | MS_REC | MS_SLAVE, NULL);
    sc_do_mount("none", dst, NULL, MS_REC | MS_SLAVE, NULL);

The first call, that establishes the bind mount, does not  use the
MS_SLAVE flag in practice. The kernel cannot change propagation and
crate a bind mount at the same time. This is documented in mount(2).

As such, remove the MS_SLAVE flag from the first line.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>