tests: add mountinfo-tool
During testing, when working with mount namespaces, we often need to ask
questions about the property of a specific mount point. Traditionally
the findmnt tool can be used for very nice, graphical exploration of the
mount namespaces. This tool has two problematic properties:

1) It doesn't show mount propagation settings with detail required for
testing. Specifically it uses custom, though arguably more readable,
vocabulary when describing propagation settings. It would be nicer to be
able to show kernel language rather than the custom language, especially
when cross-referencing kernel documentation. In addition it doesn't show
the essential property of peer group number, which is invaluable when
writing precise tests.

2) It doesn't offer unique features useful when testing across wide
range of machines. Specifically details like /dev/sda vs /dev/vda, which
are different but irrelevant to the test at hand need to be unified in
post-processing. Similarly, when tests want to measure expected
properties of a mount namespace, they will see a lot of variability
resulting from dynamic allocation of peer group numbers, block device
numbers, loopback device numbers or snap revisions. The new tool offers
a convenient option to deterministically re-number and re-name
appropriate things so that test output can be entirely deterministic if
desired.

This small python tool addresses the two shortcomings. It has convenient
syntax useful for scripting. Basic documentation and is short enough for
anyone to dive in and adjust if necessary.

It has two shortcomings that I would like to address separately:
 - doesn't have unit tests
 - it doesn't support Python 3.5 or earlier

At this time the tool is mypy and flake8 clean, though no CI is
established around that. I believe it can be addressed when unit tests
are added to the tool.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>