spread-shellcheck: add support for variants and environment
Spread support test variants where the environment section can be used
to convey specific variant information. In terms of YAML it looks like
this:

    environment:
	key/A: variant-a
	key/B: variant-b
    execute: |
	echo "$key"

Here key/A and key/B define the value of the environment variable "key"
and depending on the execution variant, the values will be "variant-a"
and "variant-b" respectively.

Currently spread-shellcheck would complain that the variable "$key" is
unset. With this patch it will know the definition of some of the
variant values. This is suboptimal but allows spread-shellcheck to not
complain about something that is clearly not broken.

As a future extension the checker could re-check the script with each
variant, carrying the exact values each time, in the hope that constant
propagation in shellcheck can make use of that information.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>