spread.yaml,tests: change MATCH and REBOOT to cmds
MATCH and REBOOT are bash functions defined by spread itself. For the
longest time we've been capturing their definitions so that scripts that
are not sourced from spread.yaml or task.yaml (so don't have access to
their definition) can still execute them.

This patch evolves the idea to unset the original definitions and
convert them to proper executables that are on PATH. This allows them to
be used from any programming language (including Python) and to be used
with the recently-introduced "not" command.

Existing users of the spread-funcs.sh are converted to just execute the
commands from PATH.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>