tests: correct application of ! / not
A single test started failing after introduction of the "not" program.
The specific code was (broken down for readability):

    su -l -c 'test-snapd-sh.with-home-plug -c \
	"ln /home/test/snap/test-snapd-sh/common/evil-2 /home/test/bin/evil-2"' test \
	2>&1 | not MATCH '.* Permission denied'

Originally there was a `!` in front of the entire expression. At runtime
we observe the actual denial, that is symlinking from
`$SNAP_USER_COMMON` to `$HOME/bin` is correctly denied, but with,
now correctly working negation, it started to fail the test.

I'm unsure as to why this happened but a working theory is that the
prior approach was designed for `set -o pipefail`, where the `su`
command would return the result of failing `ln` command which would
cause the whole pipe expression to fail.

The fix is to simply remove the negation, since the test works as
expected exactly as-is.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>