snapd: ensure GOMAXPROCS is at least 2
The Go scheduler by default has a single operating system
thread per processor core, and does it's own goroutine
scheduling inside of that thread.  For I/O operations that
the Go runtime knows about, it has mechanisms to reschedule
goroutines so the system thread isn't blocked waiting for
I/O.  If a goroutine performs a blocking system call which
the go runtime doesn't have special optimizations for, the
system thread can become blocked waiting for the syscall.
This can dramatically reduce runtime performance, and the
problem is much worse on single processor systems because
there is normally only a single system thread.

We workaround by increasing the armound of procs to a
minimum of two.