debian: fix building on eoan by tweaking golang build-deps
The go build dependencies in eoan need to be tweaked because eoan
does not ship golang-1.10 anymore. This will fix a dependency wait
in eoan.

Details for the b-d- line:
```
   golang-go (>=2:1.10)  [!powerpc] | golang-1.10 [!powerpc],
```
The first part is for 18.04+ that has go-1.10 or higher as the
default golang. The second part is for trusty,xenial where the
default golang is go-1.6 so we need to explicitly need to select
the version we want.