daemon, overlord/snapstate: support leave-cohort
With this change, you can remove (unstick?) a snap from a cohort using
the `leave-cohort` option.

An interesting thing about leave-cohort is that it's not a thing once
the request has been made: the snap setup just needs to record that it
needs to move to an empty cohort. It's a flag in daemon's
snapInstruction (and snapstate.RevisionOptions) just to not have to
fiddle with empty-string-means-no-change vs
empty-string-means-change-to-nothing.

As such, the PR might seem surprisingly simple.