packaging: fix reproducible build error
The snapd build fails to build with exactly the same hash. This
is caused by the man page embedding the date. We fix this by
setting the man-page date to the date of the debian/changelog.

Unfortunately this involves a bit of ugly sed hacking because
go-flags provides no way to change the date and faketime also
does not work on golang.