client, cmd/snap: expose the new cohort options for snap ops
This exposes the new cohort-key and leave-cohort flags for install,
refresh and switch.

Of the [Managing cohorts doc] the only bit left to do is the part
associated with `snap refresh --list`, which will probably take quite
a bit more work (because it involves a rework of how we do `snap
refresh --list`, and hopefully (!) moving to `snap refresh --dry-run`
or similar).

[Managing cohorts doc]: https://forum.snapcraft.io/t/managing-cohorts/8995