interfaces/docker-support: add controls-device-cgroup attr
This attribute allows disabling the device cgroup / udev tagging when
specified. This is because when attempting to use docker with device
specific interfaces such as gpio, or network-control, etc. that enable
a device cgroup and udev tagging, container creation fails because
dockerd isn't able to create it's own device cgroup for the container as
it is already confined by one.

A temporary remedy for this is to disable the device cgroup when needed,
by specifying this in the interface attribute.

The eventual solution of this would be for runC to learn how to nest
it's device cgroup within the one setup by snapd, but this likely won't
happen for some time, and by that point snapd will likely also have
migrated to cgroups v2 which is entirely different.

Signed-off-by: Ian Johnson <ian.johnson@canonical.com>