interfaces/system-observe: allow read on /proc/*/{cgroup,wchan}
Allow read access to /proc/<pid>/cgroup for top/ps/system monitor
utilities to also display cgroups.

Also allow read access to /proc/<pid>/wchan for showing ksym in waiting.
Note that we already allow /proc/*/stat and at column 35 the wchan field
is available to processes running with PTRACE_MODE_READ_FSCREDS which on
kernels lower than (at least) 4.4, shows the "wait channel" (ie, the
location in the kernel where the process is sleeping) as an address.
Depending on kaslr, kptr_restrict, access to /proc/kallsyms and/or
access to System.map (or from well-known distro kernels), this address
allows mapping the address to the kernel symbol (indeed, this is how ps
used to operate). 4.4 (at least) and later kernels no longer expose an
address in the wchan field since it could be used to calculate the kaslr
offset. These kernels will instead populate the wchan field with 0 or 1
depending on if the channel is blocked or not (ie, 0 if unblocked).

The 2.5 kernel introduced the /proc/<pid>/wchan file which will show the
kernel symbol in which the process is waiting without leaking addresses,
having to process the System.map or /proc/kallsyms and works regardless
of what is in /proc/<pid>/stat. Legitimate programs (eg, ps) have long
since moved to /proc/<pid>/wchan instead. Access to this file also
requires the process is running with PTRACE_MODE_READ_FSCREDS.

The manually connected system-observe interface already grants 'ptrace
(read)' for other ps accesses, so now adjust system-observe to also
allow access to /proc/<pid>/wchan.

Additionally, expand comments for 'ptrace (read)' and why system-observe
doesn't allow 'capability syslog' ir access to /proc/kallsyms, etc.