cmd/libsnap: fail early if value is NULL
When scanning for arguments of a function with variable argument list,
we are looking for key=value pairs. In the case when we saw a non-NULL
key, we must have a non-NULL value so store the search result for the
key. This mode simplifies the semantic of the function by removing an
unneeded edge case.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>