data/selinux: allow snap-confine to read symlinks on tmpfs_t
snap-confine and snap-update-ns use tmpfs when setting up the mount namespace.
When setting up the mount ns of a snap, some entries may end up with tmpfs_t
labels. We should fix it in the long run, but for now allow s-c to read
tmpfs_t (s-u-n already has the relevant permissions).

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>