cmd/libsnap: rewrite and simplify infofile query function
The use of variable length argument lists is dropped. Instead a simple
API for extracting a single key value is offered. The API is internally
powered by a scanner that does a single pass through the scanned stream,
asking a callback to inspect each key and value. Also internally there
is a private callback that looks for a specific key and stores the value,
implementing the single public API.

This method is both simpler and more extensible should we choose to
inspect more file formats and offer richer APIs later.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>