cmd/snap-confine: handle device group before pivot
Since the dawn of snapd, snap-confine would populate and join the device
cgroup nearly just before executing the target application. This is done
by interrogating udev for devices with appropriate tags and adding them
to the control group. They way they are added to the control group,
however, is pretty inefficient, as it involves executing a shell script
that writes to /sys/fs/cgroup/devices/devices.allow. The shells script
uses the interpreter #!/bin/sh forcing all base snaps to ship shell
along with the minimal set of support libraries.

Normally that was "okay" with the exception that we recently started
adding new base snaps that are otherwise empty, raising the pressure on
the need to host the spurious shell.

While a better solution would be to write to the device cgroup directly,
from snap-confine, ignoring the script, that work is somewhat larger.
As a simple alternative, just run the script before we setup and switch
to the mount namespace of the snap application.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>