cmd/snap-confine: allow opening libtinfo.so
When snap-confine runs snap-device-helper it now does so in the initial
host mount namespace. On openSUSE the shell is linked with libtinfo and
we get a denial, like this:

    type=AVC msg=audit(1561974565.075:360): apparmor="DENIED"
    operation="open" profile="/usr/lib/snapd/snap-confine"
    name="/lib64/libtinfo.so.6.1" pid=7732 comm="snap-device-hel"
    requested_mask="r" denied_mask="r" fsuid=0 ouid=0

This patch fixes that by allowing this library to be loaded. As an
alternative that was considered but discarded, udev could run before
apparmor is loaded. I think we will re-visit this once cgroup v2 work
forces changes to snap-device-helper and associated code.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>