cmd/libsnap: infofile requires trailing newlines
During code review it was raised that each line should be terminated
with the trailing newline, even the last one. While I initially objected
wishing to reuse this to more human-edited files, like /etc/os-release,
that may not fit such strict requirements, I later on agreed to change
it to ensure that an important bug fix is delivered on time.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>