tests: measure properties of various mount namespaces
The mount namespace is tricky and changes were not measured in any
explicit way, apart from what various tests were doing inside that
namespace.

This test established select measurements of the:
 - host mount namespace
 - per snap mount namespace
 - per snap, per user mount namespace

In all cases only certain distributions are measured, this is because
their initial mount namespace differs, and this echoes into the two
per-snap mount namespaces. While some careful and perhaps tricky
selective comparison might be possible I chose to instead be explicit
and only run the test on some distributions. It is always easy to
support another distribution release and to use data used by this test
to compare one distribution with another.

The set of tested distributions contains:
 - ubuntu classic 16.04 and 18.04
 - ubuntu core 16 and 18

On each distribution the set of snap mount namespace measured contains:
 - a core16-based snap (aka without specific base)
 - a core18-based snap
 - a classic snap (only on classic systems)

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>