tests: mountinfo-tool allow many --refs
When using mountinfo-tool to compute a renumbering of a per-snap mount
namespace we can use --ref to allocate numbers in a way that is comparable
with the re-numbererd host mount namespace.

To allow doing the same with the per-snap, per-user mount namespace we
need to be able to use --ref multiple times: one for the host and another
one for the per-snap numbers. This patch makes just that.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>