tests: renumber snap revisions as seen via hostfs
The mountinfo-tool program has an useful option that understands certain
properties of filesystem and allows to remove some non-determinism
caused by snap revisions. This worked fine for /snap/foo/123 and
/var/lib/snapd/snap/foo/123 but heavier use uncovered absence of support
for both of those prefixed with /var/lib/snapd/hostfs.

This patch addresses this while adding minimalistic unit tests and a way
to run them via the mountinfo-tool --self-test switch.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>