Fix error doing snap pack running failover test
It was happening that doing snap pack for core snap, sporadically the
systemd was going out of memory and hanging when running the failover
test (specially running failover test)

When using the mksquashfs command instead of snap pack the test is not
hanging anymore and the test speed is improved.

The fix was validated on gce machines and rpi as well with excelente
results.