client: add doTimeout to http.Client{Timeout}
We are currently not using a timeout AFIACT in our client.Client
code. This may lead to issues when the other side (snapd) does not
respond or is extremly slow. We got a bugreport (LP: 1837804) about
this where it happend on real devices.

We already have a a doTimeout in the client package and this PR
re-uses this and adds an integration test.