packaging/debian: don't md5sum absent files
The debian postinst hook performs some special case cleanup on an old
version of a configuration file. The logic there was written slightly
incorrectly and if the file was absent it would display a message to
stderr, like this:

    Setting up snapd (2.39.2+18.04) ...
    Installing new version of config file /etc/apparmor.d/usr.lib.snapd.snap-confine.real ...
    md5sum: /etc/apparmor.d/usr.lib.snapd.snap-confine: No such file or directory

This was never desired. This patch fixes this by checking if the file
exists in the first place and removing the useless redirect.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>