usersession: don't use a channel to signal exitOnIdle about activity.
Instead, have idleTracker provide a method that returns how long the
HTTP server has been idle (0 seconds if there are active connections).
The exitOnIdle() method can then use that duration to decide whether the
timeout has been reached, and reset the timer appropriately if it
hasn't.

As well as removing the possibility for deadlocks related to sending on
the channel, this should also make it easier to introduce more activity
tracking sources (e.g. for D-Bus).