cmd/snap-update-ns: keep track of "keep" changes
Internally snap-update-ns acts like a diff algorithm for mount
namespaces. It represents a mount namespace as a fstab-like mount
profile. Using the internal diff algorithm it decides to either remove
(unmount) add (mount) or reuse (keep) each line, which corresponds to a
mount entry.

In addition, snap-update-ns has a trespassing detection algorithm that
decides if a mount change would trespass on the mount namespace of the
host, therefore breaking the isolation and abstraction that snaps
provide. Internally snap-update-ns may decide to mount a tmpfs
somewhere, mainly to create a writable region that is isolated from the
host's mount namespace. The writable region is considered safe for
writes, because of how sharing is configured it is only visible from
the perspective of the per snap mount namespace.

When a tmpfs mounted in a prior run is reused, it would not be kept
track of for the purpose of the trespassing detector thereby causing
incorrect trespassing errors. This patch fixes that by keeping track of
"keep" changes and treating them equivalently as "mount" changes, for
the purpose of the detector.

Fixes: https://bugs.launchpad.net/snapd/+bug/1808821
Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>